legacy/requirements.txt
streamlit
streamlit-extras
ultralytics
opencv-python-headless
pillow
numpy
matplotlib
